+++ unnamed_engine/src/renderer/middleware_renderer.rs
@@ -1,8 +1,9 @@
 //! ## Middleware Renderer
 //!
 //! Defines a middleware that stores and executes everything related to the graphics library.
+use cgmath::{InnerSpace, Rotation3, Zero};
 use egui_wgpu::ScreenDescriptor;
-use wgpu::{util::DeviceExt, BindGroup, TextureFormat};
+use wgpu::{hal::auxil::db, util::DeviceExt, BindGroup, TextureFormat};
 
 use crate::gui::{egui_renderer::EguiRenderer, gui::gui};
 
@@ -49,9 +50,68 @@ const VERTICES: &[Vertex] = &[
 
 const INDICES: &[u16] = &[
   0, 1, 2, 2, 1, 3,
-  5, 4, 7, 7, 4, 6,
 ];
 
+const NUM_INSTANCES_PER_ROW: u32 = 10;
+const INSTANCE_DISPLACEMENT: cgmath::Vector3<f32> = cgmath::Vector3::new(NUM_INSTANCES_PER_ROW as f32 * 0.5, 0.0, NUM_INSTANCES_PER_ROW as f32 * 0.5);
+
+struct Instance {
+  position: cgmath::Vector3<f32>,
+  rotation: cgmath::Quaternion<f32>,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
+struct InstanceRaw {
+  model: [[f32; 4]; 4],
+}
+
+impl Instance {
+  fn to_raw(&self) -> InstanceRaw {
+    InstanceRaw {
+      model: (cgmath::Matrix4::from_translation(self.position) * cgmath::Matrix4::from(self.rotation)).into(),
+    }
+  }
+}
+
+impl InstanceRaw {
+  fn desc() -> wgpu::VertexBufferLayout<'static> {
+      use std::mem;
+      wgpu::VertexBufferLayout {
+          array_stride: mem::size_of::<InstanceRaw>() as wgpu::BufferAddress,
+          // We need to switch from using a step mode of Vertex to Instance
+          // This means that our shaders will only change to use the next
+          // instance when the shader starts processing a new instance
+          step_mode: wgpu::VertexStepMode::Instance,
+          attributes: &[
+              // A mat4 takes up 4 vertex slots as it is technically 4 vec4s. We need to define a slot
+              // for each vec4. We'll have to reassemble the mat4 in the shader.
+              wgpu::VertexAttribute {
+                  offset: 0,
+                  // While our vertex shader only uses locations 0, and 1 now, in later tutorials, we'll
+                  // be using 2, 3, and 4, for Vertex. We'll start at slot 5, not conflict with them later
+                  shader_location: 5,
+                  format: wgpu::VertexFormat::Float32x4,
+              },
+              wgpu::VertexAttribute {
+                  offset: mem::size_of::<[f32; 4]>() as wgpu::BufferAddress,
+                  shader_location: 6,
+                  format: wgpu::VertexFormat::Float32x4,
+              },
+              wgpu::VertexAttribute {
+                  offset: mem::size_of::<[f32; 8]>() as wgpu::BufferAddress,
+                  shader_location: 7,
+                  format: wgpu::VertexFormat::Float32x4,
+              },
+              wgpu::VertexAttribute {
+                  offset: mem::size_of::<[f32; 12]>() as wgpu::BufferAddress,
+                  shader_location: 8,
+                  format: wgpu::VertexFormat::Float32x4,
+              },
+          ],
+      }
+  }
+}
 
 pub struct MiddlewareRenderer {
   depth_texture: Texture,
@@ -61,6 +121,8 @@ pub struct MiddlewareRenderer {
   pipeline: wgpu::RenderPipeline,
   vertex_buffer: wgpu::Buffer,
   index_buffer: wgpu::Buffer,
+  instances: Vec<Instance>,
+  instance_buffer: wgpu::Buffer,
   num_indices: u32,
 }
 
@@ -134,7 +196,7 @@ impl MiddlewareRenderer {
       vertex: wgpu::VertexState {
         module: &shader,
         entry_point: "vs_main",
-        buffers: &[Vertex::desc()],
+        buffers: &[Vertex::desc(), InstanceRaw::desc()],
       },
       fragment: Some(wgpu::FragmentState {
         module: &shader,
@@ -183,6 +245,31 @@ impl MiddlewareRenderer {
 
     let num_indices = INDICES.len() as u32;
 
+    let instances = (0..NUM_INSTANCES_PER_ROW).flat_map(|z| {
+      (0..NUM_INSTANCES_PER_ROW).map(move |x| {
+        let position = cgmath::Vector3 {x: x as f32, y: 0.0, z: z as f32} - INSTANCE_DISPLACEMENT;
+
+        let rotation = if position.is_zero() {
+          cgmath::Quaternion::from_axis_angle(cgmath::Vector3::unit_z(), cgmath::Deg(0.0))
+        } else {
+          cgmath::Quaternion::from_axis_angle(position.normalize(), cgmath::Deg(45.0))
+        };
+
+        Instance {
+          position, rotation
+        }
+      })
+    }).collect::<Vec<_>>();
+
+    let instance_data = instances.iter().map(Instance::to_raw).collect::<Vec<_>>();
+    let instance_buffer = device.create_buffer_init(
+      &wgpu::util::BufferInitDescriptor {
+        label: Some("instance_buffer"),
+        contents: bytemuck::cast_slice(&instance_data),
+        usage: wgpu::BufferUsages::VERTEX,
+      }
+    );
+
     Self {
       depth_texture,
       diffuse_texture,
@@ -192,6 +279,8 @@ impl MiddlewareRenderer {
       vertex_buffer,
       index_buffer,
       num_indices,
+      instances,
+      instance_buffer,
     }
   }
 
@@ -247,8 +336,11 @@ impl MiddlewareRenderer {
       render_pass.set_bind_group(0, &self.diffuse_bind_group, &[]);
       render_pass.set_bind_group(1, &camera_controller.bind_group, &[]);
       render_pass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
+
+      render_pass.set_vertex_buffer(1, self.instance_buffer.slice(..));
       render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
-      render_pass.draw_indexed(0..self.num_indices, 0, 0..1);
+
+      render_pass.draw_indexed(0..self.num_indices, 0, 0..self.instances.len() as _);
     }
 
     let screen_descriptor = ScreenDescriptor {